bugfix/#91 VR 캐릭터 위젯/상호작용 정리, VR 트리거 바인딩 정리

+++ Source/Indian_Bab/Public/Widget/MainGameWidget.h
@@ -29,7 +29,7 @@ private:
 	// TObjectPtr<UEditableTextBox> Text_PlusTokenCount;
 
 	UPROPERTY(EditAnywhere)
-	float BetNum = 1;
+	int32 BetNum = 1;
 
 	UPROPERTY(meta = (BindWidget))
 	TObjectPtr<UTextBlock> Time;
